scst_user: Complete "Preparing" / "finished" symmetry
Add some TRACE statements so events sent to userland are bracketed by
"Preparing" and "finished". This makes it a little easier to find the
boundaries between the various stages of command processing in trace output.

Note, this patch does not implement a 'finished' message for TM events;
there is already a "TM reply" message that can serve that purpose.

Signed-off-by: Steven J. Magnani <steve@digidescorp.com>



git-svn-id: svn://svn.code.sf.net/p/scst/svn/trunk@5293 d57e44dd-8a1f-0410-8b47-8ef2f437770f